requirements.txt
numpy>=1.23,<3.0
Pillow>=9.0,<11.0